Going whole-hog with promises cleans stuff up
It gets rid of all the boilerplate error handling, among other things.

This also fixes a bug where the callback would never get called if
getNpmModules encountered an error.

I left in the console.log line, though it seems to me it would be
better for this module to not log anything, and let the outer layer
handle any output.